Complete landing chart link to detail page
